iot-device-dev-emulator/requirements.txt
paho-mqtt
python-dotenv
boto3
botocore